+++ src/Collector/Adapter/Iterable/ArrayObjectAdapter.php
@@ -2,11 +2,9 @@
 
 namespace FiiSoft\Jackdaw\Collector\Adapter\Iterable;
 
-use FiiSoft\Jackdaw\Collector\BaseCollector;
-use FiiSoft\Jackdaw\Collector\IterableCollector;
 use FiiSoft\Jackdaw\Stream;
 
-final class ArrayObjectAdapter extends BaseCollector implements IterableCollector, \IteratorAggregate
+final class ArrayObjectAdapter extends BaseIterableCollector
 {
     private \ArrayObject $buffer;
     
@@ -45,16 +43,6 @@ final class ArrayObjectAdapter extends BaseCollector implements IterableCollecto
     public function stream(): Stream
     {
         return Stream::from($this->buffer);
-    }
-    
-    final public function toJson(int $flags = 0): string
-    {
-        return \json_encode($this->getData(), \JSON_THROW_ON_ERROR | $flags);
-    }
-    
-    final public function toString(string $separator = ','): string
-    {
-        return \implode($separator, $this->getData());
     }
     
     public function getIterator(): \ArrayObject