+++ src/Collector/Adapter/Iterable/SplFixedArrayAdapter.php
@@ -2,11 +2,9 @@
 
 namespace FiiSoft\Jackdaw\Collector\Adapter\Iterable;
 
-use FiiSoft\Jackdaw\Collector\BaseCollector;
-use FiiSoft\Jackdaw\Collector\IterableCollector;
 use FiiSoft\Jackdaw\Stream;
 
-final class SplFixedArrayAdapter extends BaseCollector implements IterableCollector, \IteratorAggregate
+final class SplFixedArrayAdapter extends BaseIterableCollector
 {
     private \SplFixedArray $fixedArray;
     
@@ -53,16 +51,6 @@ final class SplFixedArrayAdapter extends BaseCollector implements IterableCollec
     public function stream(): Stream
     {
         return Stream::from($this->fixedArray)->notNull();
-    }
-    
-    public function toString(string $separator = ','): string
-    {
-        return \implode($separator, $this->getData());
-    }
-    
-    public function toJson(int $flags = 0): string
-    {
-        return \json_encode($this->getData(), \JSON_THROW_ON_ERROR | $flags);
     }
     
     public function getIterator()